Claude/ga4-kpi-batch/requirements.txt
google-analytics-data>=0.18.0,<1.0.0
google-api-python-client>=2.100.0,<3.0.0
google-auth>=2.20.0,<3.0.0
requests>=2.31.0,<3.0.0
